+++ egui/src/lib.rs
@@ -382,12 +382,6 @@ pub mod widgets;
 pub use epaint;
 pub use epaint::emath;
 
-// Can't add deprecation notice due to https://github.com/rust-lang/rust/issues/30827
-pub use epaint as paint; // historical reasons
-
-// Can't add deprecation notice due to https://github.com/rust-lang/rust/issues/30827
-pub use emath as math; // historical reasons
-
 pub use emath::{lerp, pos2, remap, remap_clamp, vec2, Align, Align2, NumExt, Pos2, Rect, Vec2};
 pub use epaint::{
     color, mutex,